test: rename content coverter test
