+++ src/util/__tests__/content-converter-integration-tests/mathArticle/mathArticle-test.js
@@ -15,7 +15,7 @@ import { html } from './mathArticle';
 
 const { fragment } = jsdom.JSDOM;
 
-test('serializing article with sup tag', () => {
+test('serializing article with mathml tags', () => {
   const converted = learningResourceContentToEditorValue(html, fragment);
 
   const result = learningResourceContentToHTML(converted);
